refactor: decompose map and admin layers with stronger contracts
Split map/admin responsibilities into focused modules, harden Telegram backfill pagination/auth behavior, and add runtime parsers plus targeted tests to reduce regressions and improve maintainability.
