+++ .env.example
@@ -5,3 +5,6 @@ VITE_YANDEX_METRIKA_ID=
 VITE_TELEGRAM_GEO_TTL_MINUTES=60
 VITE_TELEGRAM_TRACK_TAIL_MINUTES=30
 VITE_TELEGRAM_MAX_ACCURACY_METERS=100
+
+# Edge Function `telegram-location-bot` (supabase secrets, не Vite):
+# TELEGRAM_BACKFILL_SECRET=  # обязателен для POST .../backfill; см. README
