+++ src/admin/lib/adminApi/constants.ts
@@ -1,0 +1,1 @@
+export const PHOTOS_BUCKET = 'map-point-photos'
