+++ src/constants/layerVisibility.ts
@@ -1,0 +1,8 @@
+/** Видимость тематических слоёв карты (persist в localStorage через store). */
+export interface LayerVisibility {
+    points: boolean;
+    sockets: boolean;
+    routes: boolean;
+    bikeLanes: boolean;
+    telegramUsers: boolean;
+}
