+++ src/hooks/useLayers.ts
@@ -1,247 +1,62 @@
-import { useEffect, useState, useCallback, useRef } from 'react';
-import type { Feature, FeatureCollection, BikeLaneFeature } from '@/types/geojson';
-import type { VelojolSegment } from '@/types/velojol';
-import { fetchMapPoints, fetchMapRoutes, fetchTelegramLocations, supabase } from '@/lib/supabase';
-import {
-  mapPointsToFeatureCollection,
-  mapRoutesToFeatureCollection,
-  telegramLocationsToRecentTracksFeatureCollection,
-  telegramLocationsToUsersFeatureCollection,
-} from '@/utils/supabaseToGeojson';
-import { addLayersToMap as addLayersToMapImpl } from '@/lib/mapLayers';
-import { LAYER_IDS, type LayerKey } from '@/constants';
+import { useCallback } from 'react';
 import type { Map as MapboxMap } from 'mapbox-gl';
+import { addLayersToMap as addLayersToMapImpl } from '@/lib/mapLayers';
+import { applyVisibilityToMapLayers } from '@/constants/mapLayerRegistry';
+import type { LayerVisibility } from '@/constants/layerVisibility';
 import { useLayerVisibilityStore } from '@/hooks/useLayerVisibilityStore';
-import { useFeatureIndexes } from '@/hooks/useFeatureIndexes';
-import { useTelegramRealtime } from '@/hooks/useTelegramRealtime';
+import { useMapData } from '@/hooks/useMapData';
 
-export type { LayerKey };
-
-export interface LayerVisibility {
-  points: boolean;
-  sockets: boolean;
-  routes: boolean;
-  bikeLanes: boolean;
-  telegramUsers: boolean;
-}
-
-/** ID велодорожек, которые не показываем на карте. */
-const EXCLUDED_BIKE_LANE_IDS = new Set(['alm84', 'alm85', 'alm86', 'alm89']);
-
-function velojolToFeatureCollection(segments: VelojolSegment[]): FeatureCollection {
-  const features: BikeLaneFeature[] = segments.map((seg) => ({
-    type: 'Feature' as const,
-    geometry: {
-      type: 'LineString' as const,
-      coordinates: seg.coordinates,
-    },
-    properties: {
-      id: seg.id,
-      name: seg.name,
-      description: seg.description ?? null,
-      distance: seg.distance,
-      safetyLevel: seg.safetyLevel,
-      type: 'bikeLane' as const,
-    },
-  }));
-  return { type: 'FeatureCollection', features };
-}
+export type { LayerKey } from '@/constants';
+export type { LayerVisibility };
 
 export function useLayers() {
-  const [pointsGeo, setPointsGeo] = useState<FeatureCollection | null>(null);
-  const [routesGeo, setRoutesGeo] = useState<FeatureCollection | null>(null);
-  const [bikeLanesGeo, setBikeLanesGeo] = useState<FeatureCollection | null>(null);
-  const [telegramUsersGeo, setTelegramUsersGeo] = useState<FeatureCollection | null>(null);
-  const { visibility, toggleLayer, setLayerVisibility } = useLayerVisibilityStore();
-  const [loading, setLoading] = useState(true);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
-  const [emptyMessage, setEmptyMessage] = useState<string | null>(null);
-  const telegramRefreshSeqRef = useRef(0);
-  const refreshTimerIdRef = useRef<number | null>(null);
+    const {
+        pointsGeo,
+        routesGeo,
+        bikeLanesGeo,
+        telegramUsersGeo,
+        loading,
+        errorMessage,
+        emptyMessage,
+        getFeatureById,
+    } = useMapData();
 
-  const refreshTelegramUsers = useCallback(async () => {
-    const requestSeq = ++telegramRefreshSeqRef.current;
-    try {
-      const rows = await fetchTelegramLocations();
-      if (requestSeq !== telegramRefreshSeqRef.current) return;
-      const pointsGeo = telegramLocationsToUsersFeatureCollection(rows);
-      const tracksGeo = telegramLocationsToRecentTracksFeatureCollection(rows);
-      setTelegramUsersGeo({
-        type: 'FeatureCollection',
-        features: [...tracksGeo.features, ...pointsGeo.features],
-      });
-    } catch (error) {
-      console.error('Realtime refresh telegram users failed:', error);
-    }
-  }, []);
+    const { visibility, toggleLayer, setLayerVisibility } = useLayerVisibilityStore();
 
-  useEffect(() => {
-    const abort = { aborted: false };
-    void Promise.resolve().then(() => {
-      if (abort.aborted) return;
-      setLoading(true);
-      setErrorMessage(null);
-      setEmptyMessage(null);
-      void (async () => {
-        const [pointsResult, routesResult, telegramLocationsResult, bikeLanesModule] = await Promise.allSettled([
-          fetchMapPoints(),
-          fetchMapRoutes(),
-          fetchTelegramLocations(),
-          import('@/data/almaty.json'),
-        ]);
-        if (abort.aborted) return;
+    const addLayersToMap = useCallback(
+        (map: MapboxMap) => {
+            addLayersToMapImpl(map, {
+                pointsGeo,
+                routesGeo,
+                bikeLanesGeo,
+                telegramUsersGeo,
+                visibility,
+            });
+        },
+        [pointsGeo, routesGeo, bikeLanesGeo, telegramUsersGeo, visibility]
+    );
 
-        if (pointsResult.status === 'fulfilled') {
-          setPointsGeo(mapPointsToFeatureCollection(pointsResult.value));
-        } else {
-          setPointsGeo(null);
-        }
+    const applyVisibility = useCallback(
+        (map: MapboxMap | null) => {
+            if (!map) return;
+            applyVisibilityToMapLayers(map, visibility);
+        },
+        [visibility]
+    );
 
-        if (routesResult.status === 'fulfilled') {
-          setRoutesGeo(mapRoutesToFeatureCollection(routesResult.value));
-        } else {
-          setRoutesGeo(null);
-        }
-
-        if (telegramLocationsResult.status === 'fulfilled') {
-          const pointsGeo = telegramLocationsToUsersFeatureCollection(telegramLocationsResult.value);
-          const tracksGeo = telegramLocationsToRecentTracksFeatureCollection(telegramLocationsResult.value);
-          setTelegramUsersGeo({
-            type: 'FeatureCollection',
-            features: [...tracksGeo.features, ...pointsGeo.features],
-          });
-        } else {
-          setTelegramUsersGeo(null);
-        }
-
-        if (bikeLanesModule.status === 'fulfilled') {
-          const raw = bikeLanesModule.value.default;
-          if (Array.isArray(raw) && raw.length > 0) {
-            const segments = (raw as VelojolSegment[]).filter(
-              (seg) => !EXCLUDED_BIKE_LANE_IDS.has(seg.id)
-            );
-            setBikeLanesGeo(velojolToFeatureCollection(segments));
-          } else {
-            setBikeLanesGeo(null);
-          }
-        } else {
-          setBikeLanesGeo(null);
-        }
-
-        const errors: string[] = [];
-        if (pointsResult.status === 'rejected') {
-          const msg = pointsResult.reason instanceof Error ? pointsResult.reason.message : 'Не удалось загрузить точки.';
-          errors.push(msg);
-        }
-        if (routesResult.status === 'rejected') {
-          const msg = routesResult.reason instanceof Error ? routesResult.reason.message : 'Не удалось загрузить маршруты.';
-          errors.push(msg);
-        }
-        if (telegramLocationsResult.status === 'rejected') {
-          const msg =
-            telegramLocationsResult.reason instanceof Error
-              ? telegramLocationsResult.reason.message
-              : 'Не удалось загрузить Telegram-участников.';
-          errors.push(msg);
-        }
-        if (errors.length > 0) setErrorMessage(errors.join(' '));
-
-        const pointsCount = pointsResult.status === 'fulfilled' ? pointsResult.value.length : 0;
-        const routesCount = routesResult.status === 'fulfilled' ? routesResult.value.length : 0;
-        const telegramPointsCount =
-          telegramLocationsResult.status === 'fulfilled' ? telegramLocationsResult.value.length : 0;
-        if (pointsCount + routesCount + telegramPointsCount === 0) {
-          setEmptyMessage('Пока нет опубликованных точек, маршрутов и Telegram-локаций.');
-        }
-        setLoading(false);
-      })();
-    });
-    return () => {
-      abort.aborted = true;
-    };
-  }, []);
-
-  const scheduleRealtimeRefresh = useCallback(() => {
-    if (refreshTimerIdRef.current !== null) {
-      window.clearTimeout(refreshTimerIdRef.current);
-    }
-    refreshTimerIdRef.current = window.setTimeout(() => {
-      void refreshTelegramUsers();
-    }, 300);
-  }, [refreshTelegramUsers]);
-
-  useTelegramRealtime(supabase, scheduleRealtimeRefresh);
-
-  useEffect(() => {
-    return () => {
-      if (refreshTimerIdRef.current !== null) {
-        window.clearTimeout(refreshTimerIdRef.current);
-      }
-    };
-  }, []);
-
-  const { pointsById, routesById, bikeLanesById, telegramUsersById } = useFeatureIndexes(
-    pointsGeo,
-    routesGeo,
-    bikeLanesGeo,
-    telegramUsersGeo
-  );
-
-  const addLayersToMap = useCallback(
-    (map: MapboxMap) => {
-      addLayersToMapImpl(map, {
+    return {
         pointsGeo,
         routesGeo,
         bikeLanesGeo,
         telegramUsersGeo,
         visibility,
-      });
-    },
-    [pointsGeo, routesGeo, bikeLanesGeo, telegramUsersGeo, visibility]
-  );
-
-  const applyVisibility = useCallback(
-    (map: MapboxMap | null) => {
-      if (!map) return;
-      const setVis = (layerId: string, visible: boolean) => {
-        if (map.getLayer(layerId)) {
-          map.setLayoutProperty(layerId, 'visibility', visible ? 'visible' : 'none');
-        }
-      };
-      setVis(LAYER_IDS.points, visibility.points);
-      setVis(LAYER_IDS.sockets, visibility.sockets);
-      setVis(LAYER_IDS.routes, visibility.routes);
-      setVis(LAYER_IDS.bikeLanes, visibility.bikeLanes);
-      setVis(LAYER_IDS.telegramUsers, visibility.telegramUsers);
-      setVis(LAYER_IDS.telegramTracks, visibility.telegramUsers);
-    },
-    [visibility]
-  );
-
-  const getFeatureById = useCallback(
-    (layer: LayerKey, id: string): Feature | null => {
-      if (layer === 'points') return pointsById.get(`point:${id}`) ?? null;
-      if (layer === 'sockets') return pointsById.get(`socket:${id}`) ?? null;
-      if (layer === 'routes') return routesById.get(id) ?? null;
-      if (layer === 'bikeLanes') return bikeLanesById.get(id) ?? null;
-      return telegramUsersById.get(id) ?? null;
-    },
-    [pointsById, routesById, bikeLanesById, telegramUsersById]
-  );
-
-  return {
-    pointsGeo,
-    routesGeo,
-    bikeLanesGeo,
-    telegramUsersGeo,
-    visibility,
-    loading,
-    errorMessage,
-    emptyMessage,
-    toggleLayer,
-    setLayerVisibility,
-    addLayersToMap,
-    applyVisibility,
-    getFeatureById,
-  };
+        loading,
+        errorMessage,
+        emptyMessage,
+        toggleLayer,
+        setLayerVisibility,
+        addLayersToMap,
+        applyVisibility,
+        getFeatureById,
+    };
 }
